Created folder with CPP and moved cpp files to it.
