+++ Cpp/Brian_Kernighan’s_algorithm.cpp
