+++ Cpp/Count pairs with given sum.cpp
