+++ Cpp/Find duplicates in an array.cpp
