+++ Cpp/Fitting The Array(GeeksforGeeks).cpp
