+++ Cpp/Friendly_Pair.cpp
