+++ Cpp/Perfect Arrays(geeksforgeeks).cpp
