+++ Cpp/diserium.cpp
