+++ Cpp/palindrome.cpp
