+++ Cpp/swap_without_extra_variable.cpp
